+++ lib/twitch.dart
@@ -2,5 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/client.dart' show Twitch;
+export 'src/client.dart' show Twitch, TopGame;
 export 'src/http.dart' show TwitchHttp;
+export 'src/spec.dart' show Response, TwitchCdnImage;